requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

pyotp

kiteconnect

requests

datetime

db

azure-communication-email

azure-storage-blob

azure-identity
